3609 incorporated first pass
