+++ lambdas/dla-migration/package.json
@@ -23,9 +23,16 @@
   },
   "dependencies": {
     "@cumulus/aws-client": "18.2.0",
-    "@cumulus/logger": "18.2.0",
-    "lodash": "^4.17.21",
-    "moment": "^2.29.4",
-    "p-map": "^4.0.0"
+    "@cumulus/cumulus-message-adapter-js": "2.0.5",
+    "@cumulus/message": "18.2.0",
+    "@cumulus/types": "18.2.0",
+    "@types/minimist": "^1.2.5",
+    "lodash.clone": "^4.5.0",
+    "lodash.get": "^4.4.2",
+    "lodash.zip": "^4.2.0",
+    "minimist": "^1.2.8",
+    "p-retry": "^4.2.0",
+    "pmap": "^2.0.0",
+    "tsc": "^2.0.4"
   }
 }
