--- lambdas/update-dla-format/.gitignore
@@ -1,2 +1,0 @@
-dist/
-*.d.ts
