--- lambdas/update-dla-format/.npmignore
@@ -1,4 +1,0 @@
-/nyc.config.js
-/templates/
-/tests/
-
