--- lambdas/update-dla-format/outputs.tf
@@ -1,3 +1,0 @@
-output "function_arn" {
-  value = aws_lambda_function.update_dla_format.arn
-}
