--- lambdas/update-dla-format/package.json
@@ -1,47 +1,0 @@
-{
-  "name": "@cumulus/update-dla-format",
-  "version": "18.2.0",
-  "author": "Cumulus Authors",
-  "description": "update old DLA records to new format",
-  "license": "Apache-2.0",
-  "private": true,
-  "homepage": "https://github.com/nasa/cumulus/tree/master/lambdas/update-dla-format#readme",
-  "repository": {
-    "type": "git",
-    "url": "https://github.com/nasa/cumulus",
-    "directory": "lambdas/update-dla-format"
-  },
-  "engines": {
-    "node": ">=16.19.0"
-  },
-  "scripts": {
-    "webpack": "../../node_modules/.bin/webpack",
-    "build-lambda-zip": "(cd dist/webpack && node ../../../../bin/zip.js lambda.zip index.js migrations/*.js)",
-    "package": "npm run clean && npm run prepare && npm run webpack && npm run build-lambda-zip",
-    "tsc": "rm -rf dist && ../../node_modules/.bin/tsc",
-    "tsc:listEmittedFiles": "../../node_modules/.bin/tsc --listEmittedFiles",
-    "clean": "rm -rf node_modules && rm -rf dist",
-    "coverage": "python ../../scripts/coverage_handler/coverage.py",
-    "build": "npm run prepare",
-    "prepare": "npm run tsc",
-    "test": "../../node_modules/.bin/ava",
-    "test:coverage": "../../node_modules/.bin/nyc npm test"
-  },
-  "dependencies": {
-    "@cumulus/aws-client": "18.2.0",
-    "@cumulus/cumulus-message-adapter-js": "2.0.5",
-    "@cumulus/message": "18.2.0",
-    "@cumulus/types": "18.2.0",
-    "@types/minimist": "^1.2.5",
-    "lodash.clone": "^4.5.0",
-    "lodash.get": "^4.4.2",
-    "lodash.zip": "^4.2.0",
-    "minimist": "^1.2.8",
-    "p-retry": "^4.2.0",
-    "pmap": "^2.0.0",
-    "tsc": "^2.0.4"
-  },
-  "devDependencies": {
-    "@cumulus/common": "18.2.0"
-  }
-}
