--- lambdas/update-dla-format/tsconfig.json
@@ -1,5 +1,0 @@
-{
-  "extends": "../../tsconfig.json",
-  "include": ["./**/*.ts"],
-  "compilerOptions": { "outDir": "dist" }
-}
