--- lambdas/update-dla-format/versions.tf
@@ -1,10 +1,0 @@
-terraform {
-  required_providers {
-    aws = {
-      source = "hashicorp/aws"
-      version = "~> 5.0"
-
-    }
-  }
-  required_version = ">= 1.5"
-}
